refactor(resolver): hoist BFS state into ResolveDriver
Extract the per-resolution accumulators that used to live as ~30
function-local variables in Resolver::resolve_workspace into a
ResolveDriver<'a> struct (resolve/driver.rs). The driver borrows the
parent Resolver for its mutable cache + read hook plus read-only
configuration, while owning the BFS state: queue, resolved graph,
deferred-transitives, catalog picks, fetch scheduler, etc.

Resolver::resolve_workspace is now a 5-line dispatcher that
constructs the driver and calls run. The outer BFS loop, the
TimeBased cutoff trigger, the initial prefetch seeding, and the
per-task body are methods on ResolveDriver (bfs_loop,
seed_initial_prefetches, process_task).

In the same pass, hoist the in-flight JoinSet + primer-seeded
bookkeeping out of resolve_workspace into a FetchScheduler struct
(resolve/fetch.rs::FetchScheduler) with explicit methods
(ensure_fetch / join_next / release_in_flight / drain).

resolve.rs: 2,123 -> 161 lines. The per-task body still lives in one
method (process_task) -- breaking it into per-branch handlers
(preprocess / local-source / workspace-link / sibling-dedupe /
lockfile-reuse / fetch-and-pick) is a natural follow-up.

No behavior change; cargo test suite (~1,700 unit + integration
tests) all pass.
