+++ crates/aube-resolver/src/resolve/fetch.rs
@@ -1,31 +1,133 @@
-use crate::Error;
+use crate::{Error, FxHashSet, Resolver};
 use aube_registry::Packument;
 use aube_registry::client::RegistryClient;
 use aube_util::adaptive::AdaptiveLimit;
+use std::collections::HashSet;
 use std::path::PathBuf;
 use std::sync::Arc;
+use tokio::task::JoinSet;
+
+/// Spawns and tracks in-flight packument fetches.
+///
+/// Owns the `JoinSet` of running fetch tasks plus the bookkeeping the
+/// resolver needs to dedupe spawns (`in_flight_names`) and to know
+/// which packuments came from the bundled primer
+/// (`primer_seeded_names`, so range misses against the primer's
+/// capped history can trigger a live refetch before reporting
+/// `ERR_AUBE_NO_MATCHING_VERSION`).
+///
+/// Pre-clones the immutable Resolver bits the spawn body needs so
+/// `ensure_fetch` doesn't need a `&Resolver` borrow at call time —
+/// keeping it compatible with the BFS loop's `&mut self.resolver.cache`
+/// access pattern.
+pub(super) struct FetchScheduler {
+    in_flight: JoinSet<Result<(String, Packument, bool), Error>>,
+    in_flight_names: FxHashSet<String>,
+    primer_seeded_names: FxHashSet<String>,
+    sem: Arc<AdaptiveLimit>,
+    client: Arc<RegistryClient>,
+    cache_dir: Option<PathBuf>,
+    full_cache_dir: Option<PathBuf>,
+    mra_exclude: HashSet<String>,
+    force_metadata_primer: bool,
+    needs_time: bool,
+}
+
+pub(super) type FetchOutcome =
+    Option<Result<Result<(String, Packument, bool), Error>, tokio::task::JoinError>>;
+
+impl FetchScheduler {
+    pub(super) fn new(resolver: &Resolver, sem: Arc<AdaptiveLimit>, needs_time: bool) -> Self {
+        Self {
+            in_flight: JoinSet::new(),
+            in_flight_names: FxHashSet::default(),
+            primer_seeded_names: FxHashSet::default(),
+            sem,
+            client: resolver.client.clone(),
+            cache_dir: resolver.packument_cache_dir.clone(),
+            full_cache_dir: resolver.packument_full_cache_dir.clone(),
+            mra_exclude: resolver
+                .minimum_release_age
+                .as_ref()
+                .map(|m| m.exclude.clone())
+                .unwrap_or_default(),
+            force_metadata_primer: resolver.force_metadata_primer,
+            needs_time,
+        }
+    }
+
+    pub(super) fn in_flight_count(&self) -> usize {
+        self.in_flight.len()
+    }
+
+    /// Spawn a fetch for `name` unless one is already running for it.
+    ///
+    /// The caller is responsible for the resolver-cache gate — passing
+    /// a name that's already in the cache wastes a spawn but is
+    /// otherwise harmless.
+    pub(super) fn ensure_fetch(&mut self, name: &str, published_by: Option<&str>) {
+        if self.in_flight_names.contains(name) {
+            return;
+        }
+        self.in_flight_names.insert(name.to_string());
+        let primer_covers_cutoff = self.mra_exclude.contains(name)
+            || published_by.is_none_or(crate::primer::covers_cutoff);
+        self.in_flight.spawn(fetch_one_packument(FetchInputs {
+            name: name.to_string(),
+            client: self.client.clone(),
+            cache_dir: self.cache_dir.clone(),
+            full_cache_dir: self.full_cache_dir.clone(),
+            primer_covers_cutoff,
+            force_metadata_primer: self.force_metadata_primer,
+            sem: self.sem.clone(),
+            needs_time: self.needs_time,
+        }));
+    }
+
+    /// Wait for the next in-flight fetch to complete.
+    pub(super) async fn join_next(&mut self) -> FetchOutcome {
+        self.in_flight.join_next().await
+    }
+
+    pub(super) fn release_in_flight(&mut self, name: &str) {
+        self.in_flight_names.remove(name);
+    }
+
+    pub(super) fn note_primer_seeded(&mut self, name: String) {
+        self.primer_seeded_names.insert(name);
+    }
+
+    /// Returns true if `name` was marked as primer-seeded, removing it.
+    pub(super) fn take_primer_seeded(&mut self, name: &str) -> bool {
+        self.primer_seeded_names.remove(name)
+    }
+
+    pub(super) async fn drain(&mut self) {
+        while self.in_flight.join_next().await.is_some() {}
+    }
+}
 
 /// Inputs the packument-fetch task needs once it's spawned.
 ///
 /// All fields are owned/`Arc`-cloned so the future can be moved into
 /// the resolver's `JoinSet` without borrowing the outer scope.
-pub(super) struct FetchInputs {
-    pub(super) name: String,
-    pub(super) client: Arc<RegistryClient>,
-    pub(super) cache_dir: Option<PathBuf>,
-    pub(super) full_cache_dir: Option<PathBuf>,
+struct FetchInputs {
+    name: String,
+    client: Arc<RegistryClient>,
+    cache_dir: Option<PathBuf>,
+    full_cache_dir: Option<PathBuf>,
     /// Precomputed from the resolver's `minimum_release_age` exclude
     /// list and `published_by` cutoff — if false, the primer is
     /// bypassed even when it would otherwise be eligible.
-    pub(super) primer_covers_cutoff: bool,
+    primer_covers_cutoff: bool,
     /// `force_metadata_primer` from the resolver: when true, use the
     /// primer even for non-default registries (and rewrite tarball URLs
     /// to the active registry).
-    pub(super) force_metadata_primer: bool,
-    pub(super) sem: Arc<AdaptiveLimit>,
+    force_metadata_primer: bool,
+    sem: Arc<AdaptiveLimit>,
     /// True when the caller needs the packument's `time:` map and
     /// must therefore use the full-packument path.
-    pub(super) needs_time: bool,
+    needs_time: bool,
 }
 
 /// Body of the per-packument fetch task spawned by the resolver.
@@ -35,9 +137,7 @@ pub(super) struct FetchInputs {
 /// capped slice of high-traffic histories), so the caller knows a
 /// range miss must trigger a live registry refetch before reporting
 /// `ERR_AUBE_NO_MATCHING_VERSION`.
-pub(super) async fn fetch_one_packument(
-    inputs: FetchInputs,
-) -> Result<(String, Packument, bool), Error> {
+async fn fetch_one_packument(inputs: FetchInputs) -> Result<(String, Packument, bool), Error> {
     let FetchInputs {
         name,
         client,
